refactor: migrate Check component to new structure with rounded variant support
