+++ src/library/component/atoms/Check/Check.props.tsx
@@ -1,0 +1,8 @@
+export type Variant = 'default' | 'success' | 'info' | 'warning' | 'danger';
+export interface CheckProps {
+  checked: boolean;
+  onChange: (value: boolean) => void;
+  disabled?: boolean;
+  variant?: Variant;
+  rounded?: boolean;
+}
